rename spreadsheets_v{1,2} to workbook_json and workbook_reading
