+++ corehq/apps/app_manager/translations.py
@@ -14,7 +14,7 @@ from corehq.apps.app_manager.exceptions import (
 from corehq.apps.app_manager.models import ReportModule
 from corehq.apps.app_manager.util import save_xform
 from corehq.apps.app_manager.xform import namespaces, WrappedNode, ItextValue, ItextOutput
-from corehq.util.spreadsheets_v1.excel import HeaderValueError, WorkbookJSONReader, JSONReaderError, \
+from corehq.util.workbook_json.excel import HeaderValueError, WorkbookJSONReader, JSONReaderError, \
     InvalidExcelFileException
 
 from django.contrib import messages
