+++ corehq/apps/commtrack/tasks.py
@@ -2,7 +2,7 @@ from celery.task import task
 from corehq.apps.commtrack.consumption import recalculate_domain_consumption
 from corehq.apps.locations.bulk_management import new_locations_import
 from corehq.util.decorators import serial_task
-from corehq.util.spreadsheets_v1.excel_importer import MultiExcelImporter
+from corehq.util.workbook_json.excel_importer import MultiExcelImporter
 from django.conf import settings
 from soil import DownloadBase
