+++ corehq/apps/data_interfaces/views.py
@@ -14,7 +14,7 @@ from corehq.apps.casegroups.models import CommCareCaseGroup
 from corehq.apps.hqwebapp.templatetags.hq_shared_tags import static
 from corehq.apps.hqwebapp.utils import get_bulk_upload_form
 from corehq.apps.users.permissions import can_view_form_exports, can_view_case_exports
-from corehq.util.spreadsheets_v1.excel import JSONReaderError, WorkbookJSONReader, \
+from corehq.util.workbook_json.excel import JSONReaderError, WorkbookJSONReader, \
     InvalidExcelFileException
 from corehq.util.timezones.conversions import ServerTime
 from corehq.util.timezones.utils import get_timezone_for_user
