+++ corehq/apps/users/views/mobile/groups.py
@@ -28,7 +28,7 @@ from corehq.apps.users.forms import MultipleSelectionForm
 from corehq.apps.users.decorators import require_can_edit_commcare_users
 from corehq.apps.users.views import BaseUserSettingsView
 from corehq import privileges
-from corehq.util.spreadsheets_v1.excel import alphanumeric_sort_key
+from corehq.util.workbook_json.excel import alphanumeric_sort_key
 from dimagi.utils.decorators.memoized import memoized
