+++ custom/bihar/reports/supervisor.py
@@ -1,7 +1,7 @@
 from copy import copy
 import urllib
 from datetime import datetime, timedelta
-from corehq.util.spreadsheets_v1.excel import alphanumeric_sort_key
+from corehq.util.workbook_json.excel import alphanumeric_sort_key
 from dimagi.utils.couch.database import iter_docs
 
 from django.template.loader import render_to_string
